Split `handle` module into parts
